Review changes
- Content of Readme.md from ./samples directory moved to top level Readme.md
- Readme.md reworked
- copyright header added
- one file per type added
- #region's removed
