+++ src/devices/Mcp960X/Mcp960X.csproj
@@ -5,6 +5,8 @@
     <!--Disabling default items so samples source won't get build by the main library-->
     <EnableDefaultItems>false</EnableDefaultItems>
     <LangVersion>9</LangVersion>
+    <RootNamespace>Iot.Device.Mcp960x</RootNamespace>
+    <AssemblyName>Mcp960x</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
