+++ src/devices/Mcp960X/Register.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Iot.Device.MCP960X
+namespace Iot.Device.Mcp960x
 {
     /// <summary>
     /// Register of MCP960X
