+++ src/devices/Mcp960X/Samples/Program.cs
@@ -1,7 +1,10 @@
-﻿using System;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
 using System.Device.I2c;
 using System.Threading;
-using Iot.Device.MCP960X;
+using Iot.Device.Mcp960x;
 
 Console.WriteLine("Write I2C MCP960X - ADR 0x67 - Read ambient and hot junction temperature every 1 sec - Press Ctrl+C to end.");
 
@@ -9,7 +12,7 @@ Console.WriteLine("Write I2C MCP960X - ADR 0x67 - Read ambient and hot junction 
 // 0x67 is the device address
 I2cConnectionSettings settings = new I2cConnectionSettings(1, 0x67);
 I2cDevice i2cDevice = I2cDevice.Create(settings);
-MCP960X mcp960X = new MCP960X(i2cDevice, coldJunctionResolutionType: ColdJunctionResolutionType.N_0_25);
+Mcp960x mcp960X = new Mcp960x(i2cDevice, coldJunctionResolutionType: ColdJunctionResolutionType.N_0_25);
 
 DeviceIDType deviceIDType;
 byte major;
